+++ lib/cdt/cdt.h
@@ -133,7 +133,6 @@ struct _dtstat_s
 #define DT_BAG		0000002	/* multiset				*/
 #define DT_OSET		0000004	/* ordered set (self-adjusting tree)	*/
 #define DT_OBAG		0000010	/* ordered multiset			*/
-#define DT_STACK	0000040	/* stack: insert/delete at top		*/
 #define DT_QUEUE	0000100	/* queue: insert at top, delete at tail	*/
 #define DT_METHODS	0000377	/* all currently supported methods	*/
 
@@ -161,7 +160,6 @@ CDT_API extern Dtmethod_t* 	Dtset; ///< set with unique elements
 CDT_API extern Dtmethod_t* 	Dtbag; ///< multiset
 CDT_API extern Dtmethod_t* 	Dtoset; ///< ordered set (self-adjusting tree)
 CDT_API extern Dtmethod_t* 	Dtobag; ///< ordered multiset
-CDT_API extern Dtmethod_t*	Dtstack; ///< stack: insert/delete at top
 CDT_API extern Dtmethod_t*	Dtqueue; ///< queue: insert at top, delete at tail
 
 CDT_API extern Dtmethod_t*	Dtorder;
@@ -170,7 +168,6 @@ CDT_API extern Dtmethod_t*	Dthash;
 CDT_API extern Dtmethod_t	_Dttree;
 CDT_API extern Dtmethod_t	_Dthash;
 CDT_API extern Dtmethod_t	_Dtqueue;
-CDT_API extern Dtmethod_t	_Dtstack;
 
 CDT_API Dt_t*		dtopen(Dtdisc_t*, Dtmethod_t*);
 CDT_API int		dtclose(Dt_t*);