+++ lib/cdt/dtrenew.c
@@ -17,7 +17,7 @@ void* dtrenew(Dt_t* dt, void* obj)
 	if(!(e = dt->data->here) || _DTOBJ(e,disc->link) != obj)
 		return NULL;
 
-	if(dt->data->type&(DT_STACK|DT_QUEUE))
+	if(dt->data->type&DT_QUEUE)
 		return obj;
 	else if(dt->data->type&(DT_OSET|DT_OBAG) )
 	{	if(!e->right )	/* make left child the new root */